+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 ### next
   * tpm2_nvreadlock: long form for -P is now --auth-hierarchy
+  * tpm2_rsadecrypt: -P becomes -p
   * tpm2_nvrelease: long-form of -P becomes --auth-hierarchy
   * tpm2_nvdefine: -I becomes -p
   * tpm2_encryptdecrypt: -P becomes -p